Commented out our debugging tools

+++ Assets/Scripts/Player/PlayerMovement.cs
@@ -52,7 +52,7 @@ public class PlayerMovement : MonoBehaviour {
         else if (!cooling && col.gameObject.CompareTag("Greed"))
         {
             cooling = true;
-            int pointLoss = col.gameObject.GetComponent<parentEnemy>().damage;
+            float pointLoss = col.gameObject.GetComponent<parentEnemy>().damage;
             stats.losePoints(pointLoss);
             Invoke("stopCooling", stats.coolDownTime);
         }
